flex box on activity cards
