+++ assets/styles/activityBox.css
@@ -139,7 +139,7 @@
 /*}*/
 
 .demo-card-wide.mdl-card {
-    width: 512px;
+    width: 40em;
 }
 
 .demo-card-wide > .mdl-card__title {
